Add hartlocks
